add message new file

+++ api/controllers/AuthorController.js
@@ -4,6 +4,7 @@
  * @description :: Server-side actions for handling incoming requests.
  * @help        :: See https://sailsjs.com/docs/concepts/actions
  */
+const getMessage = sails.config.messages;
 
 module.exports = {
 
@@ -11,21 +12,23 @@ module.exports = {
     list: async (req, res) => {
         try {
             const { skip, limit,search } = req.query;
+
             let authors = await Author.find({
                 authorName: { startsWith: search } //search with authorname
-              }).skip(skip*limit).limit(limit)
+              }).skip(skip*limit)
+                .limit(limit)
                 .populate("books", { select: ['bookName'] });
 
             res.status(200).json({
-                message: "All author",
+                message: getMessage.ALL_AUTHOR,
                 Total_Author: authors.length,
                 authors
             });
 
         } catch (error) {
             console.log(error);
-            res.status(404).json({
-                message: "authors not found"
+            res.status(500).json({
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -34,6 +37,13 @@ module.exports = {
     single_author: async (req, res) => {
         try {
             const id = req.params.authorId
+
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
 
             const author = await Author.findOne({ id: id });
 
@@ -44,15 +54,14 @@ module.exports = {
                 });
             }
             else {
-                console.log(error);
                 res.status(404).json({
-                    message: "author Not Found"
+                    message: getMessage.AUTHOR_NOT_FOUND
                 });
             }
 
         } catch (error) {
             res.status(500).json({
-                error: error
+                message: getMessage.WENT_WRONG
             });
         }
 
@@ -63,20 +72,58 @@ module.exports = {
         try {
             let { authorName, authorSurname, Nationality, dateOfBirth, dateOfDeath } = req.body
 
+            let capitalize = (str) => {
+                return (
+                    str.charAt(0).toUpperCase() + str.slice(1).toLowerCase()
+                )
+            }
+
+            if (!authorName) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.AUTHOR_NAME_REQUIRE,
+                });
+            }
+            if (!authorSurname) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.AUTHOR_SURNAME_REQUIRE,
+                });
+            }
+            if (!Nationality) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.NATIONALITY_REQUIRE,
+                });
+            }
+            if (!dateOfBirth) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.DOB_REQUIRE,
+                });
+            }
+            if (!dateOfDeath) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.DOD_REQUIRE,
+                });
+            }
+
             const author = await Author.create({
-                authorName,
-                authorSurname,
+                authorName:capitalize(authorName),
+                authorSurname:capitalize(authorSurname),
                 Nationality,
                 dateOfBirth,
                 dateOfDeath
-            });
+            }).fetch()
             res.status(201).json({
-                message: "Author Create"
+                message: getMessage.AUTHOR_CREATE,
+                author
             });
 
         } catch (error) {
             res.status(500).json({
-                message: "Not Created"
+                message: getMessage.WENT_WRONG
             });
         }
 
@@ -87,26 +134,32 @@ module.exports = {
         try {
             let id = req.params.authorId;
 
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
             const Match_a = await Author.findOne({ id: id });
 
             if (Match_a) {
                 await Author.update({ id: id }).set(req.body);
 
                 res.status(200).json({
-                    message: "update author",
+                    message: getMessage.AUTHOR_UPDATE
                 });
             }
             else {
                 res.status(404).json({
-                    message: "author Not Found"
+                    message: getMessage.AUTHOR_NOT_FOUND
                 });
             }
 
 
         } catch (error) {
-            // console.log(error);
             res.status(500).json({
-                message: "Not Update"
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -116,23 +169,30 @@ module.exports = {
         try {
             let id = req.params.authorId;
 
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
             const Match_a = await Author.findOne({ id: id });
 
             if (Match_a) {
                 await Author.destroy({ id: id });
                 res.status(200).send({
-                    message: "author Delete "
+                    message: getMessage.AUTHOR_DELETE
                 });
             }
             else {
                 res.status(404).json({
-                    message: "author Not Found"
+                    message: getMessage.AUTHOR_NOT_FOUND
                 });
             }
         }
         catch (error) {
             res.status(500).json({
-                message: "Not Deleted"
+                message: getMessage.WENT_WRONG
             });
         }
     }
