+++ api/controllers/BookController.js
@@ -5,6 +5,7 @@
  * @help        :: See https://sailsjs.com/docs/concepts/actions
  */
 
+const getMessage = sails.config.messages;
 module.exports = {
 
 
@@ -12,46 +13,66 @@ module.exports = {
     list: async (req, res) => {
         try {
             const { skip, limit, search } = req.query;
-            let books = await Book.find({
-                bookName: { startsWith: search } //search with book name
-              }).skip(skip*limit).limit(limit)
+            let books = await Book.find().skip(skip * limit)
+                .limit(limit)
                 .populate("categories")
                 .populate("authors");
-            console.log(books);
+
             res.status(200).json({
-                message: "All Books",
+                message: getMessage.ALL_BOOK,
                 Total_Book: books.length,
                 Books: books
             });
+
         } catch (error) {
             console.log(error);
-            res.status(404).json({
-                message: "Not Found"
+            res.status(500).json({
+                message: getMessage.WENT_WRONG
             });
         }
     },
 
-    userList:async(req,res)=>{
+    //here user can search by book name and authorname and filter category
+    userList: async (req, res) => {
         try {
-            const{ search,author ,category}= req.query;
+            let findData = {
+                where: {
+                    bookName: "string",
+                    Price: "number"
+                }
+            }
+            let { search } = req.params
+            if (search) {
+                let searchA = search;
+                let searchB = search;
+
+                findData.where.or = [
+                    { bookName: { contains: searchA } },
+                    { Price: { contains: searchB } }
+                ]
+            }
+
+            const { searchB, searchA, category } = req.query;
+
             let books = await Book.find({
-                bookName: { startsWith: search },
-                AuthorName:{startsWith: author}
-                // categoryName:{ contains : category}
+                // bookName: { startsWith: searchB },
+                bookName: { contains: searchB },
+                Price: { contains: searchA }
             })
-            .populate("categories")
-            .populate("authors");
-         console.log(books);
-        
+                .populate("categories")
+                .populate("authors");
+
+            console.log(books);
+
             res.status(200).json({
-                count:books.length,
+                count: books.length,
                 books
             })
 
         } catch (error) {
             console.log(error);
-            res.status(404).json({
-                message: "Not Found"
+            res.status(500).json({
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -75,8 +96,8 @@ module.exports = {
             }
 
         } catch (error) {
-            res.status(404).json({
-                message: "Book Not Found"
+            res.status(500).json({
+                message: getMessage.WENT_WRONG
             });
         }
 
@@ -88,25 +109,68 @@ module.exports = {
         try {
             let { bookName, Price, publishYear, bookImage, categories, authors, users } = req.body;
 
+            let capitalize = (str) => {
+                return (
+                    str.charAt(0).toUpperCase() + str.slice(1).toLowerCase()
+                )
+            }
+
+            if (!bookName) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.BOOKNAME_REQUIRE,
+                });
+            }
+            if (!Price) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.PRICE_REQUIRE,
+                });
+            }
+            if (!publishYear) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.PUBLISHYEAR_REQUIRE,
+                });
+            }
+            if (!categories) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.CATEGORIES_REQUIRE,
+                });
+            }
+            if (!authors) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.AUTHOR_REQUIRE,
+                });
+            }
+            if (!users) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.USER_REQUIRE,
+                });
+            }
+
             const book = await Book.create({
-                bookName,
+                bookName: capitalize(bookName),
                 Price,
                 publishYear,
                 bookImage,
                 categories,
                 authors,
                 users
-            });
+            }).fetch()
 
             res.status(201).json({
-                message: "book Create",
+                message: getMessage.BOOK_CREATE,
                 Book: book
             });
 
         } catch (error) {
             console.log(error);
             res.status(500).json({
-                message: "Not Created"
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -115,51 +179,67 @@ module.exports = {
     update_book: async (req, res) => {
         try {
             let id = req.params.bookId;
+
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
             let Match_b = await Book.findOne({ id: id });
             if (Match_b) {
 
                 await Book.update({ id: id }).set(req.body);
 
                 res.status(200).json({
-                    message: "update book",
+                    message: getMessage.BOOK_UPDATE,
                 });
 
             } else {
                 res.status(404).json({
-                    message: "book not Found"
+                    message: getMessage.BOOK_NOT_FOUND
                 });
 
             }
         } catch (error) {
             res.status(500).json({
-                message: "Not Update"
+                message: getMessage.WENT_WRONG
             });
         }
 
     },
-    
+
     //delete book
     delete_book: async (req, res) => {
         try {
             let id = req.params.bookId;
+
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
             let Match_b = await Book.findOne({ id: id });
             if (Match_b) {
 
                 await Book.destroy({ id: id });
 
                 res.status(200).json({
-                    message: "Delete book",
+                    message: getMessage.BOOK_DELETE,
                 });
 
             } else {
                 res.status(404).json({
-                    message: "book Not Found"
+                    message: getMessage.BOOK_NOT_FOUND
                 });
             }
         } catch (error) {
             console.log(error);
             res.status(500).json({
-                message: "Not Delete"
+                message: getMessage.WENT_WRONG
             });
         }
     }
