+++ api/controllers/CategoryController.js
@@ -4,7 +4,7 @@
  * @description :: Server-side actions for handling incoming requests.
  * @help        :: See https://sailsjs.com/docs/concepts/actions
  */
-
+const getMessage = sails.config.messages;
 
 module.exports = {
 
@@ -12,21 +12,23 @@ module.exports = {
     list: async (req, res) => {
         try {
             const { skip, limit,search } = req.query;
+
             const category = await Category.find({
                 categoryName: { startsWith: search } //search with category name
-              }).skip(skip*limit).limit(limit)
+              }).skip(skip*limit)
+                .limit(limit)
                 .populate("books", { select: ['bookName'] });
 
             res.status(200).json({
-                message: "All Category",
+                message: getMessage.ALL_CATEGORY,
                 count:category.length,
                 category
             });
 
         } catch (error) {
             console.log(error);
-            res.status(404).json({
-                message: "Categories not found"
+            res.status(500).json({
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -34,24 +36,31 @@ module.exports = {
     //single category
     single_category: async (req, res) => {
         try {
-            const id = req.params.categoryId
+            const id = req.params.categoryId;
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
 
             const category = await Category.findOne({ id: id });
-            console.log(category);
+      
             if (category) {
                 res.status(200).json({
+                    message: getMessage.ALL_CATEGORY,
                     category: category
                 });
             }
             else {
                 res.status(404).json({
-                    message: "Category Not Found"
+                    message: getMessage.CATEGORY_NOT_FOUND
                 });
             }
 
         } catch (error) {
             res.status(500).json({
-                error: error
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -61,17 +70,32 @@ module.exports = {
         try {
             let { categoryName } = req.body;
 
+            let capitalize = (str) => {
+                return (
+                    str.charAt(0).toUpperCase() + str.slice(1).toLowerCase()
+                )
+            }
+
+            if (!categoryName) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.CATEGORY_NAME_REQUIRE,
+                });
+            }
+
             let category = await Category.create({
-                categoryName: categoryName
-            });
+                
+                categoryName: capitalize(categoryName)
+
+            }).fetch();
             res.status(201).json({
-                message: "Create Category",
+                message: getMessage.CATEGORY_CREATE,
                 category
             });
 
         } catch (error) {
             res.status(500).json({
-                message: "Not Created"
+                message: getMessage.WENT_WRONG
             });
         }
     },
@@ -82,6 +106,20 @@ module.exports = {
             let { categoryName } = req.body;
             let id = req.params.categoryId;
 
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
+            if (!categoryName) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.CATEGORY_NAME_REQUIRE,
+                });
+            }
+
             const Match = await Category.findOne({ id: id });
 
             if (Match) {
@@ -90,19 +128,19 @@ module.exports = {
                 });
 
                 res.status(200).json({
-                    message: "update Category",
+                    message: getMessage.CATEGORY_UPDATE,
                 });
             }
             else {
                 res.status(404).json({
-                    message: "Category Not Found"
+                    message: getMessage.CATEGORY_NOT_FOUND
                 });
             }
 
         } catch (error) {
             console.log(error);
             res.status(500).json({
-                message: "Not updated"
+                message: getMessage.WENT_WRONG
             });
         }
 
@@ -112,23 +150,30 @@ module.exports = {
     delete_category: async (req, res) => {
         try {
             let id = req.params.categoryId;
+            if (!id) {
+                return res.status(400).json({
+                    status: 400,
+                    message: getMessage.ID_REQUIRE,
+                });
+            }
+
             const Match = await Category.findOne({ id: id });
 
             if (Match) {
                 await Category.destroy({ id: id });
                 res.status(200).send({
-                    message: "category Delete "
+                    message: getMessage.CATEGORY_DELETE
                 });
             }
             else {
                 res.status(404).json({
-                    message: "Category Not Found"
+                    message: getMessage.CATEGORY_NOT_FOUND
                 });
             }
         }
         catch (error) {
             res.status(500).json({
-                message: "Not Deleted"
+                message: getMessage.WENT_WRONG
             });
         }
